+++ ui/main/game/galactic_war/cards/gwaio_upgrade_advancedfabricationship.js
@@ -19,7 +19,7 @@ define([
     getContext: gwoCard.getContext,
     deal: function (system, context, inventory) {
       return gwoCard.upgradeDeal(
-        gwoCard.hasUnit(inventory.units(), gwoUnit.navalFactoryAdvanced),
+        gwoCard.hasUnit(inventory.units(), gwoUnit.navalFabberAdvanced),
         gwoCard.navalWeight(inventory, 30)
       );
     },