+++ ui/main/game/galactic_war/cards/gwaio_upgrade_advancedfabricationvehicle.js
@@ -21,7 +21,7 @@ define([
     getContext: gwoCard.getContext,
     deal: function (system, context, inventory) {
       return gwoCard.upgradeDeal(
-        gwoCard.hasUnit(inventory.units(), gwoUnit.vehicleFactoryAdvanced)
+        gwoCard.hasUnit(inventory.units(), gwoUnit.vehicleFabberAdvanced)
       );
     },
     buff: function (inventory) {